requirements.txt
torch~=2.5.1
torchvision~=0.20.1
opencv-python~=4.10.0.84
Pillow~=11.0.0
numpy~=1.26.4
mediapipe~=0.10.20
matplotlib~=3.10.0